move repeater code to repeaters app
