--- corehq/apps/receiverwrapper/README.txt
@@ -1,2 +1,0 @@
-A very lightweight wrapper around the couchforms app, to add
-things like auth, etc., and provide an entrypoint.
