+++ corehq/apps/receiverwrapper/__init__.py
@@ -1,3 +1,1 @@
 from .util import submit_form_locally
-# following import registers all Generator classes to a Mapping
-import corehq.apps.receiverwrapper.repeater_generators
