+++ corehq/apps/receiverwrapper/tests/__init__.py
@@ -2,7 +2,6 @@ import logging
 from corehq.apps.receiverwrapper.util import get_version_from_appversion_text
 
 try:
-    from .test_repeater import *
     from .test_submissions import *
     from .test_submit_errors import *
     from .test_auth import *
@@ -11,7 +10,7 @@ except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
     logging.exception(e)
-    raise(e)
+    raise
 
 __test__ = {
     'get_version_from_appversion_text': get_version_from_appversion_text,
