+++ corehq/apps/repeaters/__init__.py
@@ -1,0 +1,2 @@
+# following import registers all Generator classes to a Mapping
+import corehq.apps.repeaters.repeater_generators
