+++ corehq/apps/repeaters/forms.py
@@ -1,4 +1,4 @@
-from corehq.apps.receiverwrapper.models import RegisterGenerator
+from corehq.apps.repeaters.models import RegisterGenerator
 
 from django import forms
 from django.utils.translation import ugettext_lazy as _
