+++ corehq/couchapps/__init__.py
@@ -3,6 +3,7 @@ from django.conf import settings
 
 CouchAppsPreindexPlugin.register('couchapps', __file__, {
     'form_question_schema': 'meta',
+    'receiverwrapper': 'receiverwrapper',
     'users_extra': (settings.USERS_GROUPS_DB, settings.NEW_USERS_GROUPS_DB),
     'noneulized_users': (settings.USERS_GROUPS_DB, settings.NEW_USERS_GROUPS_DB),
     'all_docs': (None, settings.NEW_USERS_GROUPS_DB, settings.NEW_FIXTURES_DB),
