+++ corehq/couchapps/receiverwrapper/indexes/search/index.js
