+++ corehq/couchapps/receiverwrapper/views/repeaters/reduce.js
